pyproject.toml
[build-system]
requires = ["flit_core >=3.2", "appdirs"]
build-backend = "flit_core.buildapi"
[project]
name = "py-oidc-auth"
description = "Pluggable OIDC/OAuth 2.1 authentication routes and token validation."
authors = [{name = "DKRZ, Clint", email = "freva@dkrz.de"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
               "Development Status :: 4 - Beta",
               "Intended Audience :: Developers",
               "License :: OSI Approved :: BSD License",
               "Operating System :: POSIX :: Linux",
               "Programming Language :: Python :: 3",
]
dynamic = ["version"]
requires-python = ">=3.8"
dependencies = [
    "httpx",
    "pydantic",
    "PyJWT[crypto]",
    "typing_extensions",
]
[project.urls]
Documentation = "https://py-oidc-auth.readthedocs.io"
Issues = "https://github.com/freva-org/py-oidc-auth/issues"
Source = "https://github.com/freva-org/py-oidc-auth/"

[project.optional-dependencies]
fastapi = ["fastapi", "python-multipart", "uvicorn"]
flask = ["flask"]
quart = ["quart"]
tornado = ["tornado"]
litestar = ["litestar"]
django = ["django>=4.0"]
dev = ["tox", "cryptography", "uvicorn", "py-oidc-auth[all,tests]"]
all = ["py-oidc-auth[fastapi,flask,quart,tornado,litestar,django]"]
tests = ["pytest",
         "pytest-env",
         "httpx",
         "mock",
         "pytest-mock",
         "pytest-cov",
         "pytest-asyncio",
         "requests",
         "uvicorn",
         ]
docs = ["sphinx-execute-code-python3",
        "sphinx-code-tabs",
        "sphinx-copybutton",
        "sphinx-sitemap",
        "sphinx-togglebutton",
        "sphinxcontrib-httpdomain",
        "sphinxext-opengraph",
        "pydata-sphinx-theme"
        ]
[package-data]
freva_auth = ["py.typed"]



[tool.pytest.ini_options]
log_cli = false
log_cli_level = "ERROR"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning:websockets",
    "ignore::DeprecationWarning:uvicorn"
]

[tool.flake8]
ignore = ["F405", "F403", "E704", "D107"]
[tool.pydocstyle]
add_ignore = ["D107"]

[tool.mypy]
files = "src"
strict = true
warn_unused_ignores = true
warn_unreachable = true
show_error_codes = true
install_types = true
non_interactive = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
ignore_missing_imports = true
allow_untyped_calls = true

[tool.isort]
known_first_party = ["oidc_auth"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]


[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise ImportError",
    "except ImportError",
]

[tool.tox]
legacy_tox_ini = """
[tox]
min_version = 4.0
env_list = docs, lint, types
passenv = *
[testenv]
passenv = *
parallel_show_output = false
[testenv:test]
description = "Run the unit tests."
deps = -e .[dev]
commands =
    python3 -m pytest -vv --cov=py_oidc_auth --cov-report=html:coverage_report --junitxml report.xml --cov-report xml tests/
    python3 -m coverage report --fail-under=98.5 --precision=2
[testenv:docs]
description = "Create the documentation."
deps = -e .[docs]
allowlist_externals = make
                      codespell
commands = codespell {posargs: --quiet-level=2} docs/source
           make -C docs clean
           make -C docs html

[testenv:lint]
description = "Check code quality."
deps = -e .[all]
       black
       isort
       flake8
       pytest-stub
       pydocstyle
       codespell
       bandit[toml]
commands =
    python3 -m isort --check --profile black -t py312 -l 79 src
    python3 -m flake8 src --count --max-complexity=10 --ignore=F405,F403,E704,W503,C901 --max-line-length=100 --statistics --show-source
    codespell {posargs: --quiet-level=2} src/
    python3 -m pydocstyle src/py_oidc_auth
    bandit -r src/py_oidc_auth --severity-level medium --confidence-level high
[testenv:types]
description = "Static type checking."
deps = .[all]
       mypy
       pytest-stub
commands = python3 -m mypy --install-types --non-interactive
[testenv:release]
deps = git-python
       packaging
       requests
       tomli
commands = python3 bump.py tag py_oidc_auth
allowlist_externals = rm
                      curl
commands_pre = curl -H 'Cache-Control: no-cache' -Ls -o bump.py https://raw.githubusercontent.com/freva-org/freva-admin/main/release.py
commands_post = rm bump.py
"""
